Moving `cursor: pointer` from dashboard-host to individual styles
Setting `cursor: pointer` makes sense for skins that also implement a
`click` handler on `body` element.

For skins that do not implement that (such as
*ets2-mobile-route-advisor*) having that style is just something that
needs to be undone.

Since the click handler is implemented individually in each skin, it
makes sense to add this style individually in each skin as well.

I have not changed the files from "source/Funbit*Server/Html", so
those have to be changed after this pull request.