Allows disabling the auto-scaling feature
Some skins (such as *ets2-mobile-route-advisor*) might want to have
total control of the viewport. For such cases, it is desirable to
disable the auto-viewport-scaling and auto-reload-when-size-changes.

This commit implements such behavior by having a non-positive width or
height.

I have not changed the files from "source/Funbit*Server/Html", so those
must be also changed after this pull request.

I have not updated the documentation, which is also another step after
this pull request.